Revert error handling (stale events are better than nothing)
This is done so empty calendars are not produced on errors (effectively
removing all events from the user's calendar). Calendar apps tend to
leave old events in place if they fail to update them (non-200 response)
and it seems like it's better to have stale events rather than nothing
while the error is being fixed.